requirements.txt
psycopg2-binary==2.9.7
paho-mqtt==1.6.1
fastapi==0.100.0
uvicorn==0.23.1
python-dotenv==1.0.0
